Add tapping navbar button on active page will scroll up; Refactor navigation to have material-top-tabs as root navigator; remove Modal; Extract navigation configurations into separate folder;
