+++ src/components/Camera/index.js
@@ -15,7 +15,7 @@ import Layout from 'constants/Layout'
 import usePrevious from 'react-use/lib/usePrevious'
 import { getCameraBonds } from 'services/Camera'
 import { BlurView } from '@react-native-community/blur'
-import Modal from 'components/Modal'
+import NativeError from 'templates/NativeError'
 import * as navigationActions from 'navigation/actions'
 
 import { withTheme } from 'react-native-paper'
@@ -116,16 +116,14 @@ const CameraComponent = ({
       />
 
       {!cameraEnabled ?
-        <View style={styling.modal}>
-          <Modal
-            cancelAction={navigationActions.navigateHome(navigation)}
-            cancelLabel={t('Cancel')}
-            confirmLabel={t('OK')}
-            confirmAction={navigationActions.navigateHome(navigation)}
-            title={t('Camera is blocked')}
-            text={t('Please enabled camera access in your phone settings to continue')}
-          />
-        </View>
+        <NativeError
+          handleCancelPress={navigationActions.navigateHome(navigation)}
+          titleText={t('Camera is blocked')}
+          messageText={t('Please enabled camera access in your phone settings to continue')}
+          actionText={t('Try again')}
+          status="failure"
+          triggerOn="failure"
+        />
       : null}
     </View>
   )
@@ -134,15 +132,6 @@ const CameraComponent = ({
 const styles = theme => StyleSheet.create({
   root: {
     flex: 1,
-  },
-  modal: {
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    right: 0,
-    bottom: 0,
-    zIndex: 1,
-    backgroundColor: theme.colors.backgroundPrimary,
   },
   cameraWrapper: {
     justifyContent: 'center',
