+++ src/navigation/AppNavigator.js
@@ -1,167 +1,40 @@
 import React from 'react'
-import { createStackNavigator, HeaderStyleInterpolators, CardStyleInterpolators } from '@react-navigation/stack'
+import { createMaterialTopTabNavigator } from '@react-navigation/material-top-tabs'
 import { withTheme } from 'react-native-paper'
-
-import TabNavigator from 'navigation/TabNavigator'
 import Layout from 'constants/Layout'
 
-import ModalScreen from 'screens/ModalScreen'
-import StoryScreen from 'screens/StoryScreen'
-import PostCreateScreen from 'screens/PostCreateScreen'
-import PostEditScreen from 'screens/PostEditScreen'
-import PostMediaScreen from 'screens/PostMediaScreen'
-import PostLikesScreen from 'screens/PostLikesScreen'
-import PostViewsScreen from 'screens/PostViewsScreen'
-import PostShareScreen from 'screens/PostShareScreen'
-import ProfileScreen from 'screens/ProfileScreen'
-import AlbumScreen from 'screens/AlbumScreen'
-import AlbumsScreen from 'screens/AlbumsScreen'
-import AlbumCreateScreen from 'screens/AlbumCreateScreen'
-import CommentsScreen from 'screens/CommentsScreen'
-import ProfileFollowedScreen from 'screens/ProfileFollowedScreen'
-import ProfileFollowerScreen from 'screens/ProfileFollowerScreen'
-import ProfileRequestsScreen from 'screens/ProfileRequestsScreen'
-import VerificationScreen from 'screens/VerificationScreen'
+import TabNavigator from 'navigation/TabNavigator'
+import CameraNavigator from 'navigation/Camera'
+import ChatNavigator from 'navigation/Chat'
 
-const AppNavigator = ({ theme }) => {
-  const Stack = createStackNavigator()
-  const stackNavigatorProps = {
-    screenOptions: {
-      headerStyleInterpolator: HeaderStyleInterpolators.forSlideUp,
-    },
-    cardStyle: {
-      backgroundColor: theme.colors.backgroundPrimary,
-    },
-  }
+const HomeNavigator = ({ theme }) => {
+  const Tab = createMaterialTopTabNavigator()
 
-  const stackScreenProps = {
-    options: {
-      headerShown: false,
-      cardStyleInterpolator: CardStyleInterpolators.forModalPresentationIOS,
-      gestureEnabled: true,
-      gestureResponseDistance: {
-        horizontal: Layout.window.width,
-        vertical: Layout.window.height,
-      },
-      gestureDirection: 'vertical',
-      cardStyle: {
-        backgroundColor: theme.colors.backgroundPrimary,
-      },
+  const tabNavigatorProps = {
+    initialRouteName: 'Feed',
+    tabBar: () => null,
+    sceneContainerStyle: {
+      backgroundColor: 'transparent',
+      width: Layout.window.width,
     },
   }
 
   return (
-    <Stack.Navigator {...stackNavigatorProps}>
-      <Stack.Screen
-        name="App"
+    <Tab.Navigator {...tabNavigatorProps}>
+      <Tab.Screen
+        name="Camera"
+        component={CameraNavigator}
+      />
+      <Tab.Screen
+        name="Feed"
         component={TabNavigator}
-        {...stackScreenProps}
       />
-
-      <Stack.Screen
-        name="Modal"
-        component={ModalScreen}
-        {...stackScreenProps}
+      <Tab.Screen
+        name="Chat"
+        component={ChatNavigator}
       />
-
-      <Stack.Screen
-        name="Story"
-        component={StoryScreen}
-        {...stackScreenProps}
-      />
-
-      <Stack.Screen
-        name="PostCreate"
-        component={PostCreateScreen}
-        {...stackScreenProps}
-      />
-
-      <Stack.Screen
-        name="PostEdit"
-        component={PostEditScreen}
-        {...stackScreenProps}
-      />
-
-      <Stack.Screen
-        name="PostMedia"
-        component={PostMediaScreen}
-        {...stackScreenProps}
-      />
-
-      <Stack.Screen
-        name="PostLikes"
-        component={PostLikesScreen}
-        {...stackScreenProps}
-      />
-
-      <Stack.Screen
-        name="PostViews"
-        component={PostViewsScreen}
-        {...stackScreenProps}
-      />
-
-      <Stack.Screen
-        name="PostShare"
-        component={PostShareScreen}
-        {...stackScreenProps}
-      />
-
-      <Stack.Screen
-        name="Profile"
-        component={ProfileScreen}
-        {...stackScreenProps}
-      />
-
-      <Stack.Screen
-        name="Album"
-        component={AlbumScreen}
-        {...stackScreenProps}
-      />
-
-      <Stack.Screen
-        name="Albums"
-        component={AlbumsScreen}
-        {...stackScreenProps}
-      />
-
-      <Stack.Screen
-        name="AlbumCreate"
-        component={AlbumCreateScreen}
-        {...stackScreenProps}
-      />
-
-      <Stack.Screen
-        name="Comments"
-        component={CommentsScreen}
-        {...stackScreenProps}
-      />
-
-      <Stack.Screen
-        name="ProfileFollowed"
-        component={ProfileFollowedScreen}
-        {...stackScreenProps}
-      />
-
-      <Stack.Screen
-        name="ProfileFollower"
-        component={ProfileFollowerScreen}
-        {...stackScreenProps}
-      />
-
-      <Stack.Screen
-        name="ProfileRequests"
-        component={ProfileRequestsScreen}
-        {...stackScreenProps}
-      />
-
-      <Stack.Screen
-        name="Verification"
-        component={VerificationScreen}
-        {...stackScreenProps}
-      />
-
-    </Stack.Navigator>
+    </Tab.Navigator>
   )
 }
 
-export default withTheme(AppNavigator)
+export default withTheme(HomeNavigator)
