+++ src/navigation/Camera/index.js
@@ -1,5 +1,24 @@
-const CameraNavigator = () => {
-  return null
+import React from 'react'
+import { createStackNavigator } from '@react-navigation/stack'
+import { withTheme } from 'react-native-paper'
+
+import * as navigationOptions from 'navigation/options'
+import CameraScreen from 'screens/CameraScreen'
+
+const CameraNavigator = ({ theme }) => {
+  const Stack = createStackNavigator()
+  const stackNavigatorDefaultProps = navigationOptions.stackNavigatorDefaultProps({ theme })
+  const stackScreenDefaultProps = navigationOptions.stackScreenDefaultProps({ theme })
+  
+  return (
+    <Stack.Navigator {...stackNavigatorDefaultProps}>
+      <Stack.Screen
+        name="Camera"
+        component={CameraScreen}
+        {...stackScreenDefaultProps}
+      />
+    </Stack.Navigator>
+  )
 }
 
-export default CameraNavigator
+export default withTheme(CameraNavigator)
